+++ lib/twenv/commands/delete_my_tweets.rb
@@ -13,6 +13,7 @@ class TWEnv::DeleteMyTweets < TWEnv::Command
   def options(slop)
     slop.on 'has-likes'     , "Only delete tweets with likes", as: :boolean, default: false
     slop.on 'has-no-likes'  , "Only delete tweets with no likes", as: :boolean, default: false
+    slop.on 'is-not-reply'  , "Only delete tweets that aren't replies", as: :boolean, default: false
     slop.on 'is-reply'      , "Only delete tweets that are replies", as: :boolean, default: false
     slop.on 'is-reply-to='  , "Only delete tweets that are a reply to the given username", as: :string, default: nil
     slop.on "has-media"     , "Only delete tweets that have media (either video or image)", default: false
@@ -46,6 +47,7 @@ class TWEnv::DeleteMyTweets < TWEnv::Command
   def filter_tweets(tweets)
     tweets = tweets.dup
     tweets.select!(&:reply?) if opts['is-reply'] || opts['is-reply-to']
+    tweets.reject!(&:reply?) if opts['is-not-reply']
     tweets.select! {|tweet| tweet.favorite_count.zero?} if opts['has-no-likes']
     tweets.select! {|tweet| tweet.favorite_count > 0 } if opts['has-likes']
     tweets.select! {|t| t.media.empty? }   if opts['no-media']